Implement resizable & keepCurrent behaviour in resizeButtonGrid
Resizable (bool) controls whether the target size window becomes resizable
by the user or not. Defaults to false.

KeepCurrent (bool), which only makes sense together with resizable=true,
makes it so changing the grid size doesn't actually change the window size
(but only the min/max sizes), so the user must manually change it to reveal
or hide content until the bounds.